+++ simulation/traffic_simulator/src/behavior/follow_trajectory.cpp
@@ -12,9 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <quaternion_operation/quaternion_operation.h>
-
 #include <arithmetic/floating_point/comparison.hpp>
+#include <geometry/quaternion/euler_to_quaternion.hpp>
+#include <geometry/quaternion/get_rotation.hpp>
+#include <geometry/quaternion/quaternion_to_euler.hpp>
 #include <geometry/vector3/hypot.hpp>
 #include <geometry/vector3/inner_product.hpp>
 #include <geometry/vector3/norm.hpp>
@@ -92,18 +93,17 @@ auto makeUpdatedStatus(
         status.lanelet_pose = lanelet_pose.value();
         const CatmullRomSpline spline(hdmap_utils->getCenterPoints(status.lanelet_pose.lanelet_id));
         const auto lanelet_quaternion = spline.getPose(status.lanelet_pose.s, true).orientation;
-        const auto lanelet_rpy =
-          quaternion_operation::convertQuaternionToEulerAngle(lanelet_quaternion);
+        const auto lanelet_rpy = math::geometry::convertQuaternionToEulerAngle(lanelet_quaternion);
         const auto entity_rpy =
-          quaternion_operation::convertQuaternionToEulerAngle(status.pose.orientation);
+          math::geometry::convertQuaternionToEulerAngle(status.pose.orientation);
         // adjust orientation in EntityStatus.pose (only pitch) and in EntityStatus.LaneletPose
-        status.pose.orientation = quaternion_operation::convertEulerAngleToQuaternion(
+        status.pose.orientation = math::geometry::convertEulerAngleToQuaternion(
           geometry_msgs::build<geometry_msgs::msg::Vector3>()
             .x(entity_rpy.x)
             .y(lanelet_rpy.y)
             .z(entity_rpy.z));
-        status.lanelet_pose.rpy = quaternion_operation::convertQuaternionToEulerAngle(
-          quaternion_operation::getRotation(lanelet_quaternion, status.pose.orientation));
+        status.lanelet_pose.rpy = math::geometry::convertQuaternionToEulerAngle(
+          math::geometry::getRotation(lanelet_quaternion, status.pose.orientation));
         status.lanelet_pose_valid = true;
       } else {
         status.lanelet_pose_valid = false;
@@ -386,7 +386,7 @@ auto makeUpdatedStatus(
                    // if entity is on lane use pitch from lanelet, otherwise use pitch on target
                    const auto pitch =
                      entity_status.lanelet_pose_valid
-                       ? -quaternion_operation::convertQuaternionToEulerAngle(
+                       ? -math::geometry::convertQuaternionToEulerAngle(
                             entity_status.pose.orientation)
                             .y
                        : std::atan2(target_position.z - position.z, std::hypot(dy, dx));
@@ -417,12 +417,10 @@ auto makeUpdatedStatus(
       desired_velocity.y, ", ", desired_velocity.z, "].");
   } else if (const auto current_velocity =
                [&]() {
-                 const auto pitch = -quaternion_operation::convertQuaternionToEulerAngle(
-                                       entity_status.pose.orientation)
-                                       .y;
-                 const auto yaw = quaternion_operation::convertQuaternionToEulerAngle(
-                                    entity_status.pose.orientation)
-                                    .z;
+                 const auto pitch =
+                   -math::geometry::convertQuaternionToEulerAngle(entity_status.pose.orientation).y;
+                 const auto yaw =
+                   math::geometry::convertQuaternionToEulerAngle(entity_status.pose.orientation).z;
                  return geometry_msgs::build<geometry_msgs::msg::Vector3>()
                    .x(std::cos(pitch) * std::cos(yaw) * speed)
                    .y(std::cos(pitch) * std::sin(yaw) * speed)
@@ -593,7 +591,7 @@ auto makeUpdatedStatus(
             .x(0.0)
             .y(std::atan2(-desired_velocity.z, std::hypot(desired_velocity.x, desired_velocity.y)))
             .z(std::atan2(desired_velocity.y, desired_velocity.x));
-        return quaternion_operation::convertEulerAngleToQuaternion(direction);
+        return math::geometry::convertEulerAngleToQuaternion(direction);
       }
     }();
 
@@ -611,7 +609,7 @@ auto makeUpdatedStatus(
     updated_status.action_status.twist.linear.z = 0;
 
     updated_status.action_status.twist.angular =
-      quaternion_operation::convertQuaternionToEulerAngle(quaternion_operation::getRotation(
+      math::geometry::convertQuaternionToEulerAngle(math::geometry::getRotation(
         entity_status.pose.orientation, updated_status.pose.orientation)) /
       step_time;